% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-curves.r
\name{gen_logistic}
\alias{gen_logistic}
\alias{gen_exponential}
\title{Curve-generating functions}
\usage{
gen_logistic(x, L, k = 1, x0 = 0)

gen_exponential(x, k, x0, offset)
}
\arguments{
\item{x}{Point at which to evaluate function}

\item{L}{Function maximum}

\item{k}{Rate parameter}

\item{x0}{Midpoint of the logistic curve}

\item{offset}{Vertical shift}
}
\description{
Curve-generating functions
}
