% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{e}
\alias{e}
\alias{rng_info}
\title{Alias and helpers for common operations}
\usage{
e()

rng_info(object)
}
\arguments{
\item{object}{Any object with an attribute named "rng_info".}
}
\description{
Alias and helpers for common operations
}
\section{Functions}{
\itemize{
\item \code{e()}: Alias for \code{\link[base:expression]{base::expression()}}

\item \code{rng_info()}: Retrieve RNG info for a simulated object

}}
