% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{num_scan_props}
\alias{num_scan_props}
\title{Distribution of the number of PET scans}
\format{
A data.table with 7 rows and 3 columns.

\itemize{
\item num_scans. \verb{<integer>} Number of PET scans.
\item N. \verb{<integer>} Number of patients in the U.C. Berkeley scan data with
n PET scans available for analysis.
\item P. \verb{<double>} The proportion of patients in the U.C. Berkeley scan data
with n PET scans available for analysis.
}
}
\usage{
num_scan_props
}
\description{
A data.table containing the distribution of the number of PET scans
available for patients in the U.C. Berkeley amyloid PET scan dataset.
}
\details{
Source: UC Berkeley - Amyloid PET 6mm Res analysis [ADNI1,GO,2,3,4]

\href{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}
}
\keyword{datasets}
