% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{scan_lag_days}
\alias{scan_lag_days}
\title{Distribution of time between scans}
\format{
An object of class \code{density} produced by \code{stats::density()}.
}
\usage{
scan_lag_days
}
\description{
The density of days since an individual's prior PET scan (for scans after the
first), based on the U.C. Berkeley amyloid PET scan dataset.
}
\details{
Source: UC Berkeley - Amyloid PET 6mm Res analysis [ADNI1,GO,2,3,4]

\href{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}
}
\keyword{datasets}
