% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-curves.r
\name{simulate_curves}
\alias{simulate_curves}
\title{Simulate amyloid curves}
\usage{
simulate_curves(
  N,
  centiloid_range = NULL,
  scan_dist = num_scan_props,
  lag_dist = scan_lag_days,
  age_scan1 = agedist_first_scan,
  age_apos_fun,
  age_apos_args,
  apos_threshold,
  genfun = NULL,
  args = NULL,
  static_vars = NULL,
  epsilon = expression(rnorm(.N, sd = 4))
)
}
\arguments{
\item{N}{Number of amyloid curves to simulate}

\item{centiloid_range}{A numeric vector of length 2 specifiying the minimum
and maximum possible centiloid values.}

\item{scan_dist}{A data.frame specifying the probability of a patient having
x PET scans.}

\item{lag_dist}{An estimated probability density (output by \code{density()})
for the number of days between two of a given individual's PET scans.}

\item{age_scan1}{A list specifying the mean and standard deviation
for the age distribution at first PET scan. E.g. \code{list(mean = 50, sd = 5)}.}

\item{age_apos_fun}{A function that generates random observations from a
distribution (e.g., rnorm, rgamma). Given as a string.}

\item{age_apos_args}{A list specifying the mean and standard deviation for the age
at which a patient's amyloid level becomes positive, based on the
centiloid threshold specified in \code{Apos_threshold}.}

\item{genfun}{A function that generates the amyloid curve.}

\item{args}{A named expression specifying the arguments to be passed to
\code{genfun}.}

\item{static_vars}{A named expression specifying input parameters to
\code{genfun()}, that remain fixed for each individual. These inputs may be
scalars or expressions specifying operations to execute in the initial
dataset created within the this function.}

\item{epsilon}{An expression that, when evaluated, will generate noise
at a given point along the amyloid curve. The resulting measurement
typically stands in as the observed amyloid level.}

\item{apos_treshhold}{The centiloid value at which a scan/individual is
considered amyloid-positive.}
}
\description{
Simulate amyloid curves
}
