% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{agedist_first_scan}
\alias{agedist_first_scan}
\title{Age distribution at first PET scan}
\format{
A data.table with 1 row and 2 columns.

\itemize{
\item mean. \verb{<double>} Mean age at first PET scan.
\item sd. \verb{<double>} Standard deviation of the age at first PET scan.
}
}
\usage{
agedist_first_scan
}
\description{
A data.table containing two columns specifying the mean age at one's first
PET scan, based on the ADNIMERGE and U.C. Berkeley scan datasets.
}
\details{
Sources:

UC Berkeley - Amyloid PET 6mm Res analysis [ADNI1,GO,2,3,4]
ADNIMERGE - Key ADNI tables merged into one table [ADNI1,GO,2,3]

\href{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}{https://ida.loni.usc.edu/explore/jsp/search/search.jsp?project=ADNI#studyFiles}
}
\keyword{datasets}
